requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
aiohttp==3.8.4
SQLAlchemy==2.0.23
alembic==1.12.1
SQLAlchemy-Utils==0.41.1
importlib-metadata~=6.0.0
python-chess==1.999
cryptography==41.0.5
pycparser==2.21
cffi==1.16.0
pytest==7.4.3
chess==1.10.0
pyhumps==3.8.0